updated tests in munging/test_dt_rows.py

+++ tests/munging/test_dt_rows.py
@@ -90,7 +90,7 @@ def test_dt0_properties(dt0):
     assert dt0.shape == (10, 3)  # must be a tuple, not a list!
     assert dt0.names == ("colA", "colB", "colC")
     assert dt0.ltypes == (ltype.bool, ltype.int, ltype.real)
-    assert dt0.stypes == (stype.bool8, stype.int16, stype.float64)
+    assert dt0.stypes == (stype.bool8, stype.int32, stype.float64)
     for i in range(dt0.ncols):
         assert frame_column_rowindex(dt0, i) is None
     frame_integrity_check(dt0)
